@@ -4,15 +4,15 @@ version: 0.1.2+3
 homepage: https://github.com/moovida/geoimage
 
 environment:
-  sdk: '>=2.8.1 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  path: ^1.7.0
+  path: ^1.8.0
 
-  dart_jts: ^0.1.0
+  dart_jts: ^0.2.0
 
   # Apache 2.0
-  image: '>=2.1.19 <=3.0.0-nullsafety.0'
+  image: ^3.0.2
     # git:
     #   url: https://github.com/brendan-duncan/image.git
     #   ref: tiff_hdr_refactor
